Refactoring functions and adding brief
