+++ Makefile
@@ -22,12 +22,13 @@ SRCS = $(SRCS_DIR)/main.c \
 		$(SRCS_DIR)/tokenize/ft_validate_syntax.c \
 		$(SRCS_DIR)/tokenize/ft_validate_syntax_utils.c \
 		$(SRCS_DIR)/tokenize/ft_validate_parentheses.c \
-		$(SRCS_DIR)/tokenize/ft_validate_parentheses_utils.c \
 		$(SRCS_DIR)/tokenize/ft_create_token_list.c \
 		$(SRCS_DIR)/tokenize/ft_create_token_list_utils.c \
+		$(SRCS_DIR)/tokenize/ft_create_token_list_utils2.c \
 		$(SRCS_DIR)/tokenize/ft_process_token_list.c \
 		$(SRCS_DIR)/tokenize/ft_process_token_list_utils.c \
 		$(SRCS_DIR)/tokenize/ft_get_wildcard_list.c \
+		$(SRCS_DIR)/tokenize/ft_get_wildcard_list_utils.c \
 		$(SRCS_DIR)/tokenize/ft_find_next_quote.c \
 		$(SRCS_DIR)/tokenize/ft_build_root.c \
 		$(SRCS_DIR)/tokenize/ft_build_tree.c \
