+++ srcs/tokenize/ft_create_token_list.c
@@ -125,7 +125,7 @@ static void	ft_process_tokens(char *s, t_list **token_list)
  * 
  * @param list A pointer to the head of the token list to validate and adjust.
  */
-void	ft_validate_export_tokens(t_list **list)
+static void	ft_validate_export_tokens(t_list **list)
 {
 	t_list	*current;
 	t_token	*token;
@@ -168,7 +168,7 @@ void	ft_validate_export_tokens(t_list **list)
  * @param list Pointer to the token list to be validated.
  * @return `true` if the tokens are valid, `false` otherwise.
  */
-bool	ft_validate_exec_tokens(t_list **list)
+static bool	ft_validate_exec_tokens(t_list **list)
 {
 	t_list	*current;
 	t_token	*token;
